// src/audio/types/node.ts
export enum NodeType {
  Analyser = 'Analyser',
  BiquadFilter = 'BiquadFilter',
  BufferSource = 'BufferSource',
  ChannelMerger = 'ChannelMerger',
  ChannelSplitter = 'ChannelSplitter',
  ConstantSource = 'ConstantSource',
  Convolver = 'Convolver',
  Delay = 'Delay',
  DynamicsCompressor = 'DynamicsCompressor',
  Gain = 'Gain',
  IIRFilter = 'IIRFilter',
  MediaElementSource = 'MediaElementSource',
  MediaStreamDestination = 'MediaStreamDestination',
  MediaStreamSource = 'MediaStreamSource',
  MediaStreamTrackSource = 'MediaStreamTrackSource',
  Oscillator = 'Oscillator',
  Panner = 'Panner',
  StereoPanner = 'StereoPanner',
  WaveShaper = 'WaveShaper'
}